+++ packages/bridge-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Uncategorized
+
+- refactor: add `.js` import extensions to Swaps-Bridge packages ([#9624](https://github.com/MetaMask/core/pull/9624))
+
 ## [78.0.0]
 
 ### Added