+++ packages/bridge-controller/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Define `QuoteResponse` v2 ([#9085](https://github.com/MetaMask/core/pull/9085))
   - `validateQuoteResponse` validates data
   - `toQuoteResponseV2` and `toQuoteResponseV1` convert quote responses to required schema when needed
+  - `toQuoteMetadataV2` and `toQuoteMetadataV1` convert quote metadata to required schema when needed
+  - `toNormalizedAmounts`converts atomic amounts to human-readable values
 - Export `BridgeAsset` and `validateBridgeAsset`, used by QuoteResponse v2, and token endpoints ([#9085](https://github.com/MetaMask/core/pull/9085))
 
 ### Changed
@@ -20,9 +22,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - convert quotes to QuoteResponse v2 in `fetchBridgeQuoteStream`
   - store quotes as QuoteResponse v2 in the BridgeController
   - `QuoteResponse` export now means v2; v1 is still exported as `QuoteResponseV1`
-  - `fetchBridgeQuoteStream` now returns `QuoteResponse` v2
+  - `fetchBridgeQuoteStream` and `fetchBatchSellTrades` now return `QuoteResponse` v2
+  - `fetchBatchSellTrades` expects V2 quotes, then transforms them to V1 for backend compatibility
   - `appendFeesToQuotes` interface now requires a chainId parameter, but still accepts both V1 and V2 quotes
 - Bump `@metamask/assets-controller` from `^11.2.0` to `^11.2.1` ([#9648](https://github.com/MetaMask/core/pull/9648))
+
+### Removed
+
+- Clean up unused utils: `isEvmQuoteResponse`, unsupporteed `XlmScope.Testnet` chainId mapping ([#9085](https://github.com/MetaMask/core/pull/9085))
 
 ## [78.0.0]